Move plate connect
