+++ ui/src/components/index.js
@@ -5,3 +5,5 @@ export { PlateSizeForm } from './activity/PlateSizeForm';
 export { Print } from './print';
 export { Editor } from './editor/Editor';
 export { PlateTabBar } from './editor/PlateTabBar';
+export { Plate } from './editor/Plate';
+export { NoPlatesMessage } from './editor/Plate';
