+++ ui/src/implementations/kapture-backend/components/index.js
@@ -6,3 +6,4 @@ export { PlateSizeForm } from '../../../components';
 export { Print } from './PrintConnected';
 export { Editor } from './EditorConnected';
 export { PlateTabBar } from './PlateTabBarConnected';
+export { Plate } from './PlateConnected';
